Refactor classroom creation page
